src/telegram-bot/requirements.txt
python-telegram-bot==22.6
openai>=1.0.0
pytz
